chore: passing tests

+++ test/auth/custom-strategy/int.spec.ts
@@ -19,7 +19,9 @@ describe('AuthStrategies', () => {
   });
 
   afterAll(async () => {
-    await payload.db.destroy(payload);
+    if (typeof payload.db.destroy === 'function') {
+      await payload.db.destroy(payload);
+    }
   });
 
   describe('create user', () => {
