+++ test/collections-graphql/int.spec.ts
@@ -17,7 +17,9 @@ describe('collections-graphql', () => {
   });
 
   afterAll(async () => {
-    await payload.db.destroy(payload);
+    if (typeof payload.db.destroy === 'function') {
+      await payload.db.destroy(payload);
+    }
   });
 
   describe('CRUD', () => {
