+++ test/relationships/int.spec.ts
@@ -32,7 +32,9 @@ describe('Relationships', () => {
   });
 
   afterAll(async () => {
-    await payload.db.destroy(payload);
+    if (typeof payload.db.destroy === 'function') {
+      await payload.db.destroy(payload);
+    }
   });
 
   beforeEach(async () => {
